Refactored FESeries to allow spacedim template for FECollection.
