+++ doc/news/changes/minor/20181017MarcFehling
@@ -1,0 +1,5 @@
+New: Classes FESeries::Fourier and FESeries::Legendre now take two
+template arguments dim and spacedim, which allow them to use corresponding
+hp::FECollection objects.
+<br>
+(Marc Fehling, 2018/10/17)
